feature: some basic data assimilation now working, temporarily created a separate pretraining cost term to cleanup the main code.. to be removed later.
